+++ contracts/handlers/oneinchV3/HOneInchV3.sol
@@ -34,6 +34,7 @@ contract HOneInchV3 is HandlerBase {
             } catch {
                 _revertMsg("swap");
             }
+            _tokenApproveZero(address(desc.srcToken), _ONEINCH_SPENDER);
         } else {
             try
                 IAggregationRouterV3(_ONEINCH_SPENDER).swap{value: desc.amount}(
@@ -70,6 +71,7 @@ contract HOneInchV3 is HandlerBase {
             // ERC20 token need to approve before unoswap
             _tokenApprove(address(srcToken), _ONEINCH_SPENDER, amount);
             returnAmount = _unoswapCall(0, data);
+            _tokenApproveZero(address(srcToken), _ONEINCH_SPENDER);
         } else {
             returnAmount = _unoswapCall(amount, data);
         }