+++ contracts/handlers/uniswapv3/HUniswapV3.sol
@@ -65,6 +65,7 @@ contract HUniswapV3 is HandlerBase {
         // Approve token
         _tokenApprove(tokenIn, address(ROUTER), params.amountIn);
         amountOut = _exactInputSingle(0, params);
+        _tokenApproveZero(tokenIn, address(ROUTER));
         WETH.withdraw(amountOut);
     }
 
@@ -88,7 +89,7 @@ contract HUniswapV3 is HandlerBase {
         // Approve token
         _tokenApprove(tokenIn, address(ROUTER), params.amountIn);
         amountOut = _exactInputSingle(0, params);
-
+        _tokenApproveZero(tokenIn, address(ROUTER));
         _updateToken(tokenOut);
     }
 
@@ -140,7 +141,7 @@ contract HUniswapV3 is HandlerBase {
         // Approve token
         _tokenApprove(tokenIn, address(ROUTER), params.amountIn);
         amountOut = _exactInput(0, params);
-
+        _tokenApproveZero(tokenIn, address(ROUTER));
         WETH.withdraw(amountOut);
     }
 
@@ -161,7 +162,7 @@ contract HUniswapV3 is HandlerBase {
         // Approve token
         _tokenApprove(tokenIn, address(ROUTER), params.amountIn);
         amountOut = _exactInput(0, params);
-
+        _tokenApproveZero(tokenIn, address(ROUTER));
         _updateToken(tokenOut);
     }
 
@@ -208,6 +209,7 @@ contract HUniswapV3 is HandlerBase {
         // Approve token
         _tokenApprove(params.tokenIn, address(ROUTER), params.amountInMaximum);
         amountIn = _exactOutputSingle(0, params);
+        _tokenApproveZero(params.tokenIn, address(ROUTER));
         WETH.withdraw(params.amountOut);
     }
 
@@ -232,7 +234,7 @@ contract HUniswapV3 is HandlerBase {
         // Approve token
         _tokenApprove(params.tokenIn, address(ROUTER), params.amountInMaximum);
         amountIn = _exactOutputSingle(0, params);
-
+        _tokenApproveZero(params.tokenIn, address(ROUTER));
         _updateToken(params.tokenOut);
     }
 
@@ -288,7 +290,7 @@ contract HUniswapV3 is HandlerBase {
         // Approve token
         _tokenApprove(tokenIn, address(ROUTER), params.amountInMaximum);
         amountIn = _exactOutput(0, params);
-
+        _tokenApproveZero(tokenIn, address(ROUTER));
         WETH.withdraw(amountOut);
     }
 
@@ -311,7 +313,7 @@ contract HUniswapV3 is HandlerBase {
         // Approve token
         _tokenApprove(tokenIn, address(ROUTER), params.amountInMaximum);
         amountIn = _exactOutput(0, params);
-
+        _tokenApproveZero(tokenIn, address(ROUTER));
         _updateToken(tokenOut);
     }